requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.17,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
celery>=5.4,<6.0
redis>=5.0,<8.0
psycopg2-binary>=2.9,<3.0
